The script divided by modules
