+++ index.html
@@ -26,6 +26,6 @@
     <div class="places-list root__section"></div>  
     <div class="popup"></div>
   </div>
-  <script src="script.js"></script>
+  <script type="module" src="./scripts/script.js"></script>
 </body>
 </html>
