+++ scripts/script.js
@@ -1,172 +1,7 @@
 /*классы*/
-class Card {
-  constructor(name, link, config) {
-    this.config = config;
-    this.cardElement = this.create(name, link);
-  }
-  create(name, link) {
-      const cardContainer = document.createElement('div');
-      const cardImage = document.createElement('div');
-      const cardIconDelete = document.createElement('button');
-      const cardDescription = document.createElement('div');
-      const cardName = document.createElement('h3');
-      const cardLikeIcon = document.createElement('button');
-      const cardLikeWrap = document.createElement('div');
-      const cardLikeCount = document.createElement('p'); 
-
-      cardContainer.classList.add(this.config.cardPlace.slice(1));
-      cardImage.classList.add(this.config.cardImageElement.slice(1));
-      cardImage.style.backgroundImage = `url(${link})`;
-      cardContainer.appendChild(cardImage);
-      cardImage.setAttribute(this.config.cardImageUrl, link);
-
-      cardIconDelete.classList.add(this.config.cardDeleteIcon.slice(1));
-      cardImage.appendChild(cardIconDelete);
-      
-      cardDescription.classList.add(this.config.cardDescription.slice(1));
-      cardContainer.appendChild(cardDescription);
-      
-      cardName.classList.add(this.config.cardTitle.slice(1));
-      cardName.textContent = name;
-      cardDescription.appendChild(cardName);
-      
-      cardLikeWrap.classList.add(this.config.cardLikeWrap.slice(1));
-      cardDescription.appendChild(cardLikeWrap);
-
-      cardLikeIcon.classList.add(this.config.cardLikeIcon.slice(1));
-      cardLikeWrap.appendChild(cardLikeIcon);
-
-      cardLikeCount.classList.add(this.config.cardLikeCount.slice(1));
-      cardLikeCount.textContent = '0';
-      cardLikeWrap.appendChild(cardLikeCount);
-
-      return cardContainer;
-  }
-  static like(element, classToToggle) {
-    element.classList.toggle(classToToggle.slice(1));
-  }
-  static remove(element, parentNode) {
-    const elementToDelete = element.closest(parentNode);
-    elementToDelete.parentNode.removeChild(elementToDelete);
-  }
-}
-
-class CardList {
-  constructor(config) {
-    this.config = config;
-    this.container = document.querySelector(config.renderContainer);
-    this.render();
-    this.container.addEventListener('click', this.clickHandler.bind(this));
-  }
-  addCard(name, link) {
-    const {cardElement} = new Card(name, link, this.config);
-    this.container.appendChild(cardElement);
-  }
-  render() {
-    const serverAPI = new Api(this.config);
-    serverAPI.loadUserInfo();
-    serverAPI.getInitialCards()
-      .then((res) => {
-          if (res) {
-              for (let i = 0; i < res.length; i++) {
-                  this.addCard(res[i].name, res[i].link);
-              }
-          } else {
-            return Promise.reject(`Нет данных`);
-          }
-      })
-      .catch((err) => {
-          console.log('Ошибка render', err);
-      });
-    }
-    clickHandler(event) {
-      if (event.target.classList.contains(this.config.cardDeleteIcon.slice(1))) { Card.remove(event.target, this.config.cardPlace); }
-      if (event.target.classList.contains(this.config.cardLikeIcon.slice(1))) { Card.like(event.target, this.config.likeClassToToggle); }
-  }
-}
-
-class Popup {
-  constructor(container) {
-    this.container = container;
-  }
-
-  open(content) {
-    this.container.insertAdjacentHTML('afterbegin', content);
-    this.container.querySelector('.popup__close').addEventListener('click', this.close.bind(this));
-    this.container.classList.add('popup_is-opened');
-  }
-
-  close() {
-    this.container.classList.remove('popup_is-opened');
-
-    while (this.container.firstChild) {
-        this.container.firstChild.remove();
-    }
-  }
-}
-
-class Api {
-  constructor({ server, cohort, token, contentType }) {
-    this.baseURL = `${server}/${cohort}`;
-    this.headers = {
-        authorization: token,
-        'Content-Type': contentType
-    };
-  }
-
-  getInitialCards() {
-    return fetch(`${this.baseURL}/cards`, {
-      headers: this.headers
-    })
-    .then((res) =>  {
-      if (res.ok) return res.json();
-    })
-    .then((result) => {
-      if (result) {
-        return result;
-        } else {
-          return Promise.reject(`Ошибка`);
-      }
-    })
-    .catch((err) => {
-      console.log('Нет соединения с сервером', err);
-    });
-  }
-  loadUserInfo() {
-    fetch(`${this.baseURL}/users/me`, {
-      headers: this.headers
-      })
-      .then((res) =>  {
-        if (res.ok) return res.json();
-      })
-      .then((result) => {
-        if (result) {
-          document.querySelector('.user-info__photo').style.backgroundImage = `url(${result.avatar})`;
-          document.querySelector('.user-info__name').textContent = result.name;
-          document.querySelector('.user-info__job').textContent = result.about;
-        } else {
-          return Promise.reject(`Ошибка`);
-        }
-      })
-      .catch((err) => {
-        console.log('Соединение установлено, но данные пользователя не получены', err);
-      });
-  }
-  editUserInfo(name, about) {
-    return fetch(`${this.baseURL}/users/me`, {
-      method: 'PATCH',
-      headers: this.headers,
-      body: JSON.stringify({name, about})
-    });
-  }
-  addCardOnServer(name, link) {
-    return fetch(`${this.baseURL}/cards`, {
-      method: 'POST',
-      headers: this.headers,
-      body: JSON.stringify({name,link})
-    });
-  }
-}
+import CardList from './cardlist.js';  
+import {Popup} from './popup.js';
+import {Api} from './api.js';
 
 /*переменные*/
 const config = {
